Rename utils functions
